+++ packages/runtime/src/templates/server.ts
@@ -1,6 +1,7 @@
 // eslint-disable-next-line n/no-deprecated-api -- this is what Next.js uses as well
 import { parse } from 'url'
 
+import { NextConfig } from 'next'
 import type { PrerenderManifest } from 'next/dist/build'
 import type { BaseNextResponse } from 'next/dist/server/base-http'
 import type { NodeRequestHandler, Options } from 'next/dist/server/next-server'
@@ -12,6 +13,7 @@ import {
   localizeRoute,
   localizeDataRoute,
   unlocalizeRoute,
+  getMatchedRoute,
 } from './handlerUtils'
 
 interface NetlifyConfig {
@@ -23,6 +25,10 @@ const getNetlifyNextServer = (NextServer: NextServerType) => {
   class NetlifyNextServer extends NextServer {
     private netlifyConfig: NetlifyConfig
     private netlifyPrerenderManifest: PrerenderManifest
+
+    public getAppRouterReactVersion(): string {
+      return this.nextConfig.experimental?.serverActions ? 'experimental' : 'next'
+    }
 
     public constructor(options: Options, netlifyConfig: NetlifyConfig) {
       super(options)
@@ -47,7 +53,7 @@ const getNetlifyNextServer = (NextServer: NextServerType) => {
         const { url, headers } = req
 
         // conditionally use the prebundled React module
-        this.netlifyPrebundleReact(url)
+        this.netlifyPrebundleReact(url, this.nextConfig, parsedUrl)
 
         // intercept on-demand revalidation requests and handle with the Netlify API
         if (headers['x-prerender-revalidate'] && this.netlifyConfig.revalidateToken) {
@@ -76,24 +82,20 @@ const getNetlifyNextServer = (NextServer: NextServerType) => {
     }
 
     // doing what they do in https://github.com/vercel/vercel/blob/1663db7ca34d3dd99b57994f801fb30b72fbd2f3/packages/next/src/server-build.ts#L576-L580
-    private netlifyPrebundleReact(path: string) {
+    private async netlifyPrebundleReact(path: string, { basePath, trailingSlash }: NextConfig, parsedUrl) {
       const routesManifest = this.getRoutesManifest?.()
       const appPathsRoutes = this.getAppPathRoutes?.()
-
       const routes = routesManifest && [...routesManifest.staticRoutes, ...routesManifest.dynamicRoutes]
-      const matchedRoute = routes?.find((route) => new RegExp(route.regex).test(new URL(path, 'http://n').pathname))
+      const matchedRoute = await getMatchedRoute(path, routes, parsedUrl, basePath, trailingSlash)
       const isAppRoute = appPathsRoutes && matchedRoute ? appPathsRoutes[matchedRoute.page] : false
 
       if (isAppRoute) {
         // app routes should use prebundled React
         // eslint-disable-next-line no-underscore-dangle
-        process.env.__NEXT_PRIVATE_PREBUNDLED_REACT = this.nextConfig.experimental?.serverActions
-          ? 'experimental'
-          : 'next'
+        process.env.__NEXT_PRIVATE_PREBUNDLED_REACT = this.getAppRouterReactVersion()
+
         return
       }
-
-      // pages routes should use use node_modules React
       // eslint-disable-next-line no-underscore-dangle
       process.env.__NEXT_PRIVATE_PREBUNDLED_REACT = ''
     }